+++ Changelog.md
@@ -1,4 +1,7 @@
 # Changelog
+ 
+## v1.3a
+* Combined all scripts into one file
 
 ## v1.2a
 * Respawnable_on_call.add should be called respawnable_on_call.new -- Fixed